Moved XYVector and include local headers as strings
