+++ src/lib/snake.h
@@ -2,20 +2,13 @@
 #define SNAKE_H_
 
 #include <stdlib.h>
-
-struct XYVector {
-    int x;
-    int y;
-};
+#include "vector.h"
 
 struct Snake {
     struct XYVector position;
     struct Snake *next;
 };
 
-struct XYVector generateXyVector(int x, int y);
-struct XYVector addXyVectors(struct XYVector vector1, struct XYVector vector2);
-int equalXyVectors(struct XYVector vector1, struct XYVector vector2);
 void freeSnake(struct Snake **snake);
 void feedSnake(struct Snake **snake);
 void moveSnake(struct Snake **snake, struct XYVector vector);
